[scaffold-lint-rule] Refactor script structure
