+++ packages/eslint-plugin-wonder-blocks/scripts/generate-lint-rule/__tests__/templates.test.ts
