+++ packages/eslint-plugin-wonder-blocks/scripts/generate-lint-rule/patches.ts
