+++ packages/eslint-plugin-wonder-blocks/scripts/generate-lint-rule/templates.ts
